pyproject.toml
[project]
name = "okagents"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "camel-ai==0.2.22",
    "neo4j==5.28.1",
    "pre-commit>=4.1.0",
    "ruff==0.7.4",
]

[tool.hatch.build.targets.wheel]
packages = ["okagents"]

[tool.jupyter]
notebook-dir = "okagents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 79
skip-string-normalization = true

[tool.ruff]
target-version = "py39"
line-length = 119
indent-width = 4

[tool.ruff.lint]
ignore = ["C901", "C403", "C419", "E501", "E722", "E741", "W605", "C408"]
select = ["C", "E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-third-party = ["torch", "transformers", "vllm"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
